+++ src/components/Footer/Footer.js
@@ -1,10 +1,26 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
 import FooterLi from './footerLi/footerLi';
-import FOOTER_DATA from './footerData';
 import '../Footer/Footer.scss';
 
 const Footer = () => {
+  const FOOTER_DATA = [
+    {
+      id: 1,
+      data: '협력업체등록',
+    },
+    { id: 2, data: '공지사항' },
+    {
+      id: 3,
+      data: '고객센터',
+    },
+    { id: 4, data: '개인정보처리방침' },
+    {
+      id: 5,
+      data: '이용약관',
+    },
+  ];
+
   return (
     <footer className="footer">
       <div className="footMain inner">
@@ -26,8 +42,8 @@ const Footer = () => {
           </p>
         </div>
         <div className="sns">
-          <Link className="footerInsta" to="/" />
-          <Link className="footerFace" to="/" />
+          <a className="footerInsta" href="/" />
+          <a className="footerFace" href="/" />
         </div>
       </div>
     </footer>
