--- src/components/Footer/footerData.js
@@ -1,18 +1,0 @@
-const FOOTER_DATA = [
-  {
-    id: 1,
-    data: '협력업체등록',
-  },
-  { id: 2, data: '공지사항' },
-  {
-    id: 3,
-    data: '고객센터',
-  },
-  { id: 4, data: '개인정보처리방침' },
-  {
-    id: 5,
-    data: '이용약관',
-  },
-];
-
-export default FOOTER_DATA;
